moved each instruction handler out of switch stmt and into separate functions

+++ include/ace-vm/Array.hpp
@@ -11,14 +11,14 @@ namespace vm {
 
 class Array {
 public:
-    Array(int size = 0);
+    Array(size_t size = 0);
     Array(const Array &other);
     ~Array();
 
     Array &operator=(const Array &other);
     inline bool operator==(const Array &other) const { return this == &other; }
 
-    inline int GetSize() const { return m_size; }
+    inline size_t GetSize() const { return m_size; }
     inline Value &AtIndex(int index) { return m_buffer[index]; }
     inline const Value &AtIndex(int index) const { return m_buffer[index]; }
     inline void AtIndex(int index, const Value &value) { m_buffer[index] = value; }
@@ -30,8 +30,8 @@ public:
     void GetRepresentation(utf::Utf8String &out_str, bool add_type_name = true) const;
 
 private:
-    int m_size;
-    int m_capacity;
+    size_t m_size;
+    size_t m_capacity;
     Value *m_buffer;
 };
